+++ src/dmLib/__init__.py
@@ -1,6 +1,6 @@
 """
-Desin Margins Library
----------------------
+Design Margins Library
+----------------------
 
 A library for computing design margins using 
 fuzzy logic and probabilistic methods for 
@@ -8,11 +8,12 @@ uncertainty modelling
 
 """
 
-__version__ = '0.3.0'
-__all__=["triangularFunc", "fuzzySet", "fuzzyRule", 
+__version__ = '0.4.0'
+__all__=["triangularFunc", "fuzzySet", "fuzzyRule",
     "fuzzySystem", "gaussianFunc", "Distribution"
-    "Design"]
+    "Design", "MarginNode"]
 
 from .fuzzyLib import triangularFunc, fuzzySet, fuzzyRule, fuzzySystem
 from .uncertaintyLib import gaussianFunc, Distribution
 from .DOELib import Design
+from .designMarginsLib import MarginNode